# swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: Collecto API
  description: API de autenticación y gestión de sesión para la plataforma Collecto
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/collecto/Collecto/1.0.0
  - url: http://localhost:3000
    description: Servidor de desarrollo local
paths:
...
  /api/adverts:
    get:
      tags:
        - Anuncios
      summary: Obtener todos los anuncios (con paginación y filtros)
      description: Obtiene una lista de anuncios con opción de paginación, filtros y el estado de favoritos si el usuario está autenticado
      operationId: getAllAdverts
      parameters:
        - name: page
          in: query
          description: Página de resultados (por defecto 1)
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Número máximo de resultados por página (por defecto 10)
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          description: Criterio de ordenación (por defecto "createdAt")
          schema:
            type: string
            default: "createdAt"
      responses:
        '200':
          description: Lista de anuncios con estado de favoritos (si el usuario está autenticado)
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    description: Total de anuncios disponibles
                  adverts:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: ID del anuncio
                        title:
                          type: string
                          description: Título del anuncio
                        description:
                          type: string
                          description: Descripción del anuncio
                        price:
                          type: number
                          format: double
                          description: Precio del artículo
                        isFavorite:
                          type: boolean
                          description: Indica si el anuncio es favorito del usuario autenticado
        '404':
          description: No se encontraron anuncios
        '500':
          description: Error interno en el servidor
  /api/adverts/{slug}:
    get:
      tags:
        - Anuncios
      summary: Obtiene los detalles de un anuncio por slug
      description: Devuelve un anuncio específico utilizando su slug
      operationId: getAdvertBySlug
      parameters:
        - name: slug
          in: path
          required: true
          description: Slug único del anuncio
          schema:
            type: string
      responses:
        '200':
          description: Detalles del anuncio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Advert'
        '404':
          description: Anuncio no encontrado
  /api/adverts/search:
    get:
      tags:
        - Anuncios
      summary: Buscar anuncios con filtros
      description: Permite buscar anuncios con diversos filtros como título, precio, estado, etc. Si el usuario está autenticado, incluye el estado de favoritos
      operationId: searchAdverts
      parameters:
        - name: title
          in: query
          description: Título del anuncio (filtro)
          schema:
            type: string
        - name: priceMin
          in: query
          description: Precio mínimo del artículo
          schema:
            type: number
            format: double
        - name: priceMax
          in: query
          description: Precio máximo del artículo
          schema:
            type: number
            format: double
        - name: tags
          in: query
          description: Filtra anuncios por etiquetas (separadas por coma)
          schema:
            type: string
        - name: status
          in: query
          description: Estado del anuncio (disponible, reservado, vendido)
          schema:
            type: string
        - name: transaction
          in: query
          description: Tipo de transacción (compra o venta)
          schema:
            type: string
        - name: collection
          in: query
          description: Filtra por colección del anuncio
          schema:
            type: string
        - name: createdAtMin
          in: query
          description: Fecha de creación mínima
          schema:
            type: string
            format: date
        - name: createdAtMax
          in: query
          description: Fecha de creación máxima
          schema:
            type: string
            format: date
        - name: brand
          in: query
          description: Marca del producto
          schema:
            type: string
        - name: product_type
          in: query
          description: Tipo de producto
          schema:
            type: string
        - name: universe
          in: query
          description: Universo del artículo (por ejemplo, Marvel, Star Wars, etc.)
          schema:
            type: string
        - name: condition
          in: query
          description: Condición del artículo (nuevo, usado, etc.)
          schema:
            type: string
        - name: slug
          in: query
          description: Filtro por slug del anuncio
          schema:
            type: string
        - name: page
          in: query
          description: Página de resultados (por defecto 1)
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Número máximo de resultados por página (por defecto 10)
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          description: Criterio de ordenación (por defecto "createdAt")
          schema:
            type: string
            default: "createdAt"
        - name: sortOrder
          in: query
          description: Orden de clasificación (-1 para descendente, 1 para ascendente)
          schema:
            type: integer
            default: -1
      responses:
        '200':
          description: Lista de anuncios con estado de favoritos si el usuario está autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  adverts:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: ID del anuncio
                        title:
                          type: string
                          description: Título del anuncio
                        description:
                          type: string
                          description: Descripción del anuncio
                        price:
                          type: number
                          format: double
                          description: Precio del artículo
                        isFavorite:
                          type: boolean
                          description: Indica si el anuncio es favorito del usuario autenticado
        '404':
          description: No se encontraron anuncios
        '500':
          description: Error interno en el servidor
...
  /api/adverts/{id}/status:
    patch:
      tags:
        - Anuncios
      summary: Cambiar el estado y visibilidad de un anuncio
      description: Permite cambiar el estado del anuncio (disponible, reservado, vendido) y su visibilidad
      operationId: updateAdvertStatus
      parameters:
        - name: id
          in: path
          required: true
          description: ID del anuncio
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - disponible
                    - reservado
                    - vendido
      responses:
        '200':
          description: Estado del anuncio actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "El estado del anuncio ha sido cambiado a vendido"
                  advert:
                    $ref: '#/components/schemas/Advert'
        '404':
          description: Anuncio no encontrado
        '403':
          description: No tienes permiso para editar este anuncio
  /api/adverts/{id}/picture:
    post:
      tags:
        - Anuncios
      summary: Subir imágenes para un anuncio
      description: Permite subir imágenes para un anuncio
      operationId: uploadImages
      parameters:
        - name: id
          in: path
          required: true
          description: ID del anuncio
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                images:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Imágenes subidas correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Imágenes subidas correctamente"
                  images:
                    type: array
                    items:
                      type: string
        '404':
          description: Anuncio no encontrado
  /api/adverts/{id}/picturee:
    get:
      tags:
        - Anuncios
      summary: Obtener imágenes de un anuncio
      description: Devuelve todas las imágenes asociadas a un anuncio
      operationId: getImages
      parameters:
        - name: id
          in: path
          required: true
          description: ID del anuncio
          schema:
            type: string
      responses:
        '200':
          description: Imágenes del anuncio
          content:
            application/json:
              schema:
                type: object
                properties:
                  images:
                    type: array
                    items:
                      type: string
                      description: URL de la imagen
        '404':
          description: Anuncio no encontrado
...
  /api/users/me/adverts:
    get:
      tags:
        - Anuncios
      summary: Ver anuncios de uno mismo
      description: Obtiene los anuncios del usuario autenticado. Incluye el total de anuncios publicados.
      operationId: getOwnAdverts
      responses:
        '200':
          description: Lista de anuncios del usuario autenticado, incluyendo el total de anuncios
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalAdverts:
                    type: integer
                    description: Total de anuncios publicados por el usuario
                  adverts:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: ID del anuncio
                        title:
                          type: string
                          description: Título del anuncio
                        description:
                          type: string
                          description: Descripción del anuncio
                        price:
                          type: number
                          format: double
                          description: Precio del artículo
        '404':
          description: Usuario no encontrado o sin anuncios
        '500':
          description: Error interno en el servidor
  /api/advrets:
    post:
      tags:
        - Usuarios
      summary: Crear un nuevo anuncio
      description: Crea un anuncio para el usuario autenticado
      operationId: createAdvert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                price:
                  type: number
                  format: double
                transaction:
                  type: string
                  enum:
                    - venta
                    - compra
                status:
                  type: string
                  enum:
                    - disponible
                    - reservado
                    - vendido
                product_type:
                  type: string
                universe:
                  type: string
                condition:
                  type: string
                  enum:
                    - nuevo
                    - usado
                    - roto
                collection:
                  type: string
                brand:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Anuncio creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Anuncio creado"
                  advert:
                    $ref: '#/components/schemas/Advert'
        '400':
          description: Faltan campos obligatorios
  /api/adverts/{id}:
    get:
      tags:
        - Usuarios
      summary: Obtener un anuncio específico
      description: Devuelve los detalles de un anuncio específico
      operationId: getAdvertById
      parameters:
        - name: id
          in: path
          required: true
          description: ID del anuncio
          schema:
            type: string
      responses:
        '200':
          description: Detalles del anuncio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Advert'
        '404':
          description: Anuncio no encontrado
  /api/adveerts/{id}:
    put:
      tags:
        - Usuarios
      summary: Editar un anuncio
      description: Permite editar un anuncio creado por el usuario autenticado
      operationId: editAdvert
      parameters:
        - name: id
          in: path
          required: true
          description: ID del anuncio
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                price:
                  type: number
                  format: double
                transaction:
                  type: string
                  enum:
                    - venta
                    - compra
                status:
                  type: string
                  enum:
                    - disponible
                    - reservado
                    - vendido
                product_type:
                  type: string
                universe:
                  type: string
                condition:
                  type: string
                  enum:
                    - nuevo
                    - usado
                    - roto
                collection:
                  type: string
                brand:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Anuncio actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Anuncio actualizado"
                  advert:
                    $ref: '#/components/schemas/Advert'
        '404':
          description: Anuncio no encontrado
        '403':
          description: No tienes permiso para editar este anuncio
  /api/advertes/{id}:
    delete:
      tags:
        - Usuarios
      summary: Eliminar un anuncio
      description: Elimina un anuncio creado por el usuario autenticado
      operationId: deleteAdvert
      parameters:
        - name: id
          in: path
          required: true
          description: ID del anuncio
          schema:
            type: string
      responses:
        '200':
          description: Anuncio eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Anuncio eliminado"
        '404':
          description: Anuncio no encontrado
        '403':
          description: No tienes permiso para eliminar este anuncio
...
components:
  schemas:
...
    Advert:
      type: object
      properties:
        id:
          type: string
          description: ID del anuncio
        title:
          type: string
          description: Título del anuncio
        description:
          type: string
          description: Descripción del anuncio
        price:
          type: number
          format: double
          description: Precio del artículo
        transaction:
          type: string
          description: Tipo de transacción (compra o venta)
        status:
          type: string
          description: Estado del anuncio (disponible, reservado, vendido)
        images:
          type: array
          items:
            type: string
            description: URL de la imagen del anuncio
        createdAt:
          type: string
          format: date-time
          description: Fecha de creación del anuncio
        updatedAt:
          type: string
          format: date-time
          description: Fecha de la última actualización del anuncio
...